update/fix : Change the filledForms to flagfilled boolean and makng corresponding chnages and changing all forms to a single form
