--- src/app/Diet/diet-info-form/page.jsx
@@ -1,12 +1,0 @@
-import DietQuestionnaireCarousel from "@/components/forms/DietInfoForm";
-import React from "react";
-
-function DietInfoPage() {
-  return (
-    <div className="bg-primary">
-      <DietQuestionnaireCarousel></DietQuestionnaireCarousel>
-    </div>
-  );
-}
-
-export default DietInfoPage;
