--- src/app/Gym/workout-info-form/page.jsx
@@ -1,12 +1,0 @@
-import WorkoutQuestionnaireCarousel from "@/components/forms/WorkoutInfoForm";
-import React from "react";
-
-function WorkoutInfoPage() {
-  return (
-    <div>
-      <WorkoutQuestionnaireCarousel></WorkoutQuestionnaireCarousel>
-    </div>
-  );
-}
-
-export default WorkoutInfoPage;
