+++ src/lib/FormQuestions.js
@@ -1,0 +1,125 @@
+const questions = [
+  {
+    id: "name",
+    question: "What is your name?",
+    type: "text",
+    description: "Please enter your full name",
+  },
+  {
+    id: "age",
+    question: "What is your age?",
+    type: "number",
+    description: "Enter your age in years",
+  },
+  {
+    id: "height",
+    question: "What is your height?",
+    type: "number",
+    description: "Enter your height in centimeters",
+  },
+  {
+    id: "weight",
+    question: "What is your weight?",
+    type: "number",
+    description: "Enter your weight in kilograms",
+  },
+  {
+    id: "fitnessGoal",
+    question: "What is your fitness goal?",
+    type: "select",
+    options: ["Weight Loss", "Muscle Gain", "General Fitness"],
+    description: "Choose the primary goal for your fitness journey",
+  },
+  {
+    id: "frequency",
+    question: "How many days a week can you exercise?",
+    type: "select",
+    options: ["1-2", "3-4", "5+"],
+    description: "Select the number of days you can dedicate to working out",
+  },
+  {
+    id: "healthConditions",
+    question: "Any health conditions, physical limitations or injuries?",
+    type: "text",
+    description: "Mention any health concerns that may affect your workout",
+    placeholder: "If yes, please specify",
+  },
+  {
+    id: "dietPreference",
+    question: "Do you have any dietary Preferences?",
+    type: "select",
+    options: [
+      "None",
+      "Vegetarian",
+      "Vegan",
+      "Gluten-free",
+      "Lactose-free",
+      "Other",
+    ],
+    description: "Select your dietary restrictions if any, or select 'None'",
+  },
+  {
+    id: "region",
+    question: "What is your preferred cuisine?",
+    type: "select",
+    options: [
+      "INDIAN",
+      "AMERICAN",
+      "CHINESE",
+      "ITALIAN",
+      "MEXICAN",
+      "CONTINENTAL",
+      "OTHER",
+    ],
+  },
+  {
+    id: "fitnessGoal",
+    question: "What is your primary fitness goal?",
+    type: "select",
+    options: [
+      "Lose Weight",
+      "Build Muscle",
+      "Improve Endurance",
+      "Overall Health",
+    ],
+  },
+  {
+    id: "frequency",
+    question: "How often do you currently exercise?",
+    type: "select",
+    options: ["Never", "1-2 times/week", "3-4 times/week", "5+ times/week"],
+  },
+  {
+    id: "ExerciseType",
+    question: "What type of workouts do you prefer?",
+    type: "select",
+    options: [
+      "Cardio",
+      "StrengthTraining",
+      "HIIT",
+      "Yoga",
+      "Mixed",
+      "MuscleBuilding",
+    ],
+  },
+  {
+    id: "fitnessLevel",
+    question: "How would you rate your current fitness level?",
+    type: "select",
+    options: ["SKINNY", "FIT", "FAT", "OBESE"],
+  },
+  {
+    id: "intensityLevel",
+    question: "What is your preferred workout intensity level?",
+    type: "select",
+    options: ["LOW", "MEDIUM", "HIGH"],
+  },
+  {
+    id: "equipmentAccess",
+    question: "Do you have access to a gym or home equipment?",
+    type: "select",
+    options: ["Gym", "HomeEquipment", "Both", "Neither"],
+  },
+];
+
+export default questions;
